New CLI and a couple new features
One thing I added was some general info about the file formats in BinderTool/equipParamWeaponInfo.txt. This may not be the best place for it but I don't have anywhere else to put it (there seems to be essentially no ER modding/reversing/data mining community yet), so I'm just gonna leave it here for now.